Processing file uploads
- Refactor code

#2353
